More than 4 years later, updated Roots.php
This was one of the worst chunks of code I've ever written - factoring numbers by iterating in multiple places with tons of while loops? Outstanding move by 15 year old me. I finally found some free time to at least clear up this part of the overall not-so-good code, hope it hasn't influenced anyone, ever. This code as a whole is still messy, still isn't the best and there are many things I wouldn't do these days. But it was uploaded for archival purposes anyway.